🚚 템플릿 레이어 패키지 통합.
굳이 나눌 필요가 없네.
