+++ ui/src/commonMain/kotlin/kr/lul/stringnotebook/ui/template/ListHeader.kt
@@ -1,10 +1,10 @@
-package kr.lul.stringnotebook.ui.template.notebook
+package kr.lul.stringnotebook.ui.template
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import kr.lul.stringnotebook.state.organism.notebook.NewNotebookHeaderButton
-import kr.lul.stringnotebook.state.template.notebook.NotebookListHandler
+import kr.lul.stringnotebook.state.template.NotebookListHandler
 import kr.lul.stringnotebook.ui.organism.notebook.ListHeaderButton
 import kotlin.uuid.ExperimentalUuidApi
